requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
drf-spectacular==0.26.5
Pillow==9.5.0
python-decouple==3.8
psycopg2-binary==2.9.9
django-filter==23.4
django-modeltranslation==0.18.11
coverage==7.3.2
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
